TE-001 Fixed data types, spacing and unnecessary elements
